Finished first pass at command line interface and moved functions to timesheet
